server_py/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
requests==2.32.3
pydantic==2.10.6
python-dotenv==1.0.1

# Social / RSS / scheduling dependencies
feedparser==6.0.10
tweepy==4.14.0
googlesearch-python==1.2.3
schedule==1.2.0
youtube-transcript-api==0.6.1

